// Final Project - Final version with TypeScript version/crwn-clothing/src/Store/user/user.types.ts
export enum USER_ACTION_TYPE {
  Switch_Current_User = "user/Switch_Current_User",
  Check_User_Session = "user/Check_User_Session",
  Google_Sign_In_Start = "user/Google_Sign_In_Start",
  Email_Sign_In_Start = "user/Email_Sign_In_Start",
  Sign_In_Success = "user/Sign_In_Success",
  Sign_In_Falied = "user/Sign_In_Falied",
  Sign_Up_Start = "user/Sign_Up_Start",
  Sign_Up_Success = "user/Sign_Up_Success",
  Sign_Up_Failed = "user/Sign_Up_Failed",
  Sign_Out_Start = "user/Sign_Out_Start",
  Sign_Out_Success = "user/Sign_Out_Success",
  Sign_Out_Failed = "user/Sign_Out_Failed",
}